backend/requirements.txt
abstract_singleton==1.0.1
aiohttp==3.8.4
aiohttp-retry==2.9.1
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
beautifulsoup4==4.12.2
black==23.3.0
blessed==1.21.0
blis==0.7.9
browser-cookie3==0.20.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2025.4.26
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.71
colorama==0.4.6
commonmark==0.9.1
confection==0.0.4
coverage==7.2.5
cryptography==40.0.2
cssselect==1.2.0
cymem==2.0.7
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
fastapi==0.115.12
filelock==3.12.0
flake8==6.0.0
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
identify==2.5.23
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
langcodes==3.3.0
loguru==0.7.0
lxml==5.4.0
lz4==4.4.4
Markdown==3.3.7
MarkupSafe==2.1.2
mccabe==0.7.0
measurement==3.2.2
mergedeep==1.3.4
mkdocs==1.4.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
myfitnesspal==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
orjson==3.8.11
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pathy==0.10.1
pdfkit==1.0.0
Pillow==9.5.0
pinecone-client==2.2.1
pip-system-certs==4.0
platformdirs==3.5.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
preshed==3.0.8
protobuf==4.22.3
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.23.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==9.11
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0
pyyaml_env_tag==0.1
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
reportlab==4.4.1
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==12.6.0
rsa==4.9
selenium==4.1.4
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.46.2
sympy==1.14.0
thinc==8.1.9
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
twilio==9.6.1
typer==0.7.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
uvicorn==0.34.2
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.13
webdriver-manager==3.8.6
websocket-client==1.5.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
